+++ src/style_helpers.rs
@@ -45,6 +45,12 @@ mod repeat_fn_tests {
     }
 }
 
+#[cfg(feature = "grid")]
+/// Returns a grid template containing `count` evenly sized tracks
+pub fn evenly_sized_tracks(count: u16) -> Vec<TrackSizingFunction> {
+    vec![repeat(count, vec![flex(1.0)])]
+}
+
 /// Specifies a grid line to place a grid item between in CSS Grid Line coordinates:
 ///  - Positive indicies count upwards from the start (top or left) of the explicit grid
 ///  - Negative indicies count downwards from the end (bottom or right) of the explicit grid